changed type of seed from float to int

+++ xaitk_saliency/impls/perturb_image/rise.py
@@ -40,7 +40,7 @@ class RISEPertubation (PerturbImage):
         N: int,
         s: int,
         p1: float,
-        seed: Optional[float] = None
+        seed: Optional[int] = None
     ):
         """
         Generate a set of random binary masks
